==> 00_Stress_dict.ipynb <==
# Generating Spanish phonetic correspondence with prosody

The syltippy package (https://github.com/nur-ag/syltippy) will be used to generate syllabified (stress-indicated) outputs for each word found in the transcriptions. Then, the corresponding vowels in the TextGrid-formant dataframes will be marked as either stressed or unstressed.

import pandas as pd
import csv
import re
import numpy as np
from syltippy import syllabize

# create function to take in word input and generate uppercase for stressed syllable

def stress(word):
    syllables, stress = syllabize(word)
    return ','.join(s if stress != i else s.upper() for (i, s) in enumerate(syllables))

# enter word
stress("guerra")

vowels = pd.read_csv("data/allvowels.csv")
vowels.head()

# apply lambda function of 'stress' on the 'word' column

vowels["stress"] = vowels["word"].apply(lambda x : stress(str(x)))
vowels.sample(5)

# create separate column to hold only the vowels in each syllable

vowels["syll_vowels"] = vowels["stress"].apply(lambda x: re.sub(r'[^,aeiouAEIOUáéíóúÁÉÍÓÚ]', '', x))
vowels.sample(5)

# define function to return index of 'vowels' column with stress

def is_stress(word):
    # convert to list
    word = word.split(",")
    stress_vowel = 0
    for syllable in word:
        if syllable.isupper():
            stress_vowel = word.index(syllable)
    return stress_vowel

is_stress("e,I")

# create new column which gives vowel number in given word that has stress
vowels["stress_vowel"] = vowels["syll_vowels"].apply(lambda x: is_stress(x))
print(len(vowels))
vowels.sample(5)

Now all that's left to do is mark each vowel in the textgrid dataframe according to its index in the word it is produced in. Then, for each vowel index that matches the index given in `stress_vowel`, the vowel will be marked as `stressed`.

vowels["vowel_ind"] = vowels.groupby(["Participant", "t1_wd"]).cumcount()
vowels = vowels.reset_index(drop = True)
print(len(vowels))
vowels.head(5)

vowels["stress"] = np.where(vowels['stress_vowel'] == vowels['vowel_ind'], "stressed", "unstressed")
print(len(vowels))
vowels.head(5)

Now, create a function that will take as an input a dictionary dataframe with columns `word` and `ipa` and will also take as input a dataframe of vowel productions (from formants and TextGrid), also with a column `word` and columns `Participant`, `t1_wd`, and `t1_ph`. The below function requires that the formant dataframe only include vowel formants (i.e. non vowels are exluded).

def get_stress(vowels):
    # import required packages
    import pandas as pd
    import csv
    import re
    import numpy as np
    from syltippy import syllabize
    
    # def fxn to create stress column in dictionary
    def stress(word):
        syllables, stress = syllabize(word)
        return ','.join(s if stress != i else s.upper() for (i, s) in enumerate(syllables))
    
    # add column to dictionary
    vowels["stress_syll"] = vowels["word"].apply(lambda x : stress(str(x)))
    
    # create separate column to hold only the vowels in each word
    vowels["syll_vowels"] = vowels["stress_syll"].apply(lambda x: re.sub(r'[^,aeiouAEIOUáéíóúÁÉÍÓÚ]', '', x))
    
    # define function to return index of 'vowels' column with stress
    def is_stress(word):
        # convert to list
        word = word.split(",")
        stress_vowel = 0
        for syllable in word:
            if syllable.isupper():
                stress_vowel = word.index(syllable)
        return stress_vowel
    
    # create new column which gives vowel number in given word that has stress
    vowels["stress_vowel"] = vowels["syll_vowels"].apply(lambda x: is_stress(x))
    
    # determine index of vowel in df
    vowels["vowel_ind"] = vowels.groupby(["Participant", "t1_wd"]).cumcount()
    vowels = vowels.reset_index(drop = True)
    
    # add column to formants to indicate stress
    vowels["stress"] = np.where(vowels['stress_vowel'] == vowels['vowel_ind'], "stressed", "unstressed")
    
    # drop unnecesary columns
    vowels = vowels.drop(["syll_vowels", "stress_syll"], axis = 1)
       
    return vowels
