==> 02_formants_tg.ipynb <==
# Create .csv for CBAS word list data

Dataframes for female and male data are concatenated to yield a single dataframe of word list elicitations from CBAS.

import pandas as pd
import os
from audiolabel import read_label

## Import textgrids

cbasdf = pd.DataFrame({
    'relpath': 'textgrids/cbas',
    'fname': ['p112.TextGrid',
              'p119.TextGrid',
              'p113.TextGrid',
              'p115.TextGrid',
              'p120.TextGrid',
              'p124.TextGrid'],
    'subject': ['p112', 'p119', 'p113', 'p115', 'p120', 'p124']
})

dimedf = pd.DataFrame({
    'relpath': 'textgrids/dime',
    'fname' : os.listdir("textgrids/dime")})
dimedf['subject'] = dimedf['fname'].apply(lambda x: x[:4])

tgdf = pd.concat([cbasdf, dimedf], ignore_index = True)

# inputs 

def tg2df(row):
    '''Load 'phone' and 'word' tiers from a textgrid and merge them.
    
    Parameters
    ----------
    
    row: named tuple
    A namedtuple as provided by `itertuples` that can be used to load a Praat
    textgrid from a path identified by row.relpath and row.fname. The textgrid is
    expected to have 'phone' and 'word' tiers.

    Returns
    -------
    
    mergedf: the merged dataframe.
    '''
    [wddf, phdf] = read_label(
        os.path.join(row.relpath, row.fname).replace("\\","/"),
        ftype='praat',
        tiers=['word', 'phone']
    )
    # Throw an error if tiers are not strictly hierarchical.
    # words contain phones
    assert(wddf.t1.isin(phdf.t1).all())
    assert(wddf.t2.isin(phdf.t2).all())
    
    # Add phone duration and speaker
    phdf['dur_ph'] = phdf.t2 - phdf.t1
    phdf['Participant'] = row.subject

    # Merge phone and word tiers.
    phwddf = pd.merge_asof(
        phdf.rename({'t1': 't1_ph', 't2': 't2_ph'}, axis='columns'),
        wddf.drop('fname', axis='columns') \
            .rename({'t1': 't1_wd', 't2': 't2_wd'}, axis='columns'),
        left_on='t1_ph',
        right_on='t1_wd'
    )

    # Add word-init and -final columns
    phwddf['is_wdinit_ph'] = phwddf.t1_ph == phwddf.t1_wd
    phwddf['is_wdfin_ph'] = phwddf.t2_ph == phwddf.t2_wd

    # Merge context tier and return the result.
    return phwddf

dflist = [tg2df(row) for row in tgdf.itertuples()]

alldf = pd.concat(dflist, ignore_index=True)

alldf.sample(10)

Create cols `prev_ph` and `next_ph` containing previous and following phones.

alldf['prev_ph'] = alldf.phone.shift(1).fillna('')
alldf['next_ph'] = alldf.phone.shift(-1).fillna('')
alldf = alldf[alldf["phone"]!=""]
alldf = alldf.reset_index(drop = True)

alldf

# save df

alldf.to_csv("data/all_phones.csv", index = False)

## Import formant data

# import cbas female
cbas_fem = pd.read_csv("data/cbas_female_formants.txt", sep = "\t")
cbas_fem = cbas_fem.rename(columns = {"Filename": "Participant",
                           "Segment label":"phone"})
cbas_fem["Gender"] = "Female"
cbas_fem["Corpus"] = "CBAS"

# import cbas male
cbas_male = pd.read_csv("data/cbas_male_formants.txt", sep = "\t")
cbas_male = cbas_male.rename(columns = {"Filename": "Participant",
                           "Segment label":"phone"})
cbas_male["Gender"] = "Male"
cbas_male["Corpus"] = "CBAS"

# combine cbas female and male
cbas = pd.concat([cbas_male, cbas_fem], ignore_index = True)
cbas.head()

# import dime female
dime_fem = pd.read_csv("data/dime_female_formants.txt", sep = "\t")
dime_fem = dime_fem.rename(columns = {"Filename": "Participant",
                           "Segment label":"phone"})
dime_fem["Gender"] = "Female"
dime_fem["Corpus"] = "DIMEx100"

# import dime male
dime_male = pd.read_csv("data/dime_male_formants.txt", sep = "\t")
dime_male = dime_male.rename(columns = {"Filename": "Participant",
                           "Segment label":"phone"})
dime_male["Gender"] = "Male"
dime_male["Corpus"] = "DIMEx100"

dime = pd.concat([dime_male, dime_fem], ignore_index = True)
dime.head()

# combine cbas and dime, female and male
formants = pd.concat([cbas, dime], ignore_index = True)
formants

## Merge formant and textgrid data

data = formants.merge(alldf, left_index = True, right_index = True, how = "outer")
data = data.drop(["Participant_y", "phone_y"], axis = 1)
data = data.rename(columns = {"Participant_x": "Participant",
                             "phone_x": "phone"})
data.sample(10)

import re

# fix phones from txt file, remove + following some vowels
data["phone"] = data["phone"].apply(lambda x: re.sub("([aeiou])\+", "\1", x))

# replace `r(` with `rf` for consistency
data['word'] = data['word'].apply(lambda x: re.sub("r\(", "rf", x))

# fix notation in dimex corpus, where V_7 yields accented V
data['word'] = data['word'].apply(lambda x: re.sub("a_7", "á", x))
data['word'] = data['word'].apply(lambda x: re.sub("i_7", "í", x))
data['word'] = data['word'].apply(lambda x: re.sub("o_7", "ó", x))
data['word'] = data['word'].apply(lambda x: re.sub("u_7", "ú", x))
data['word'] = data['word'].apply(lambda x: re.sub("e_7", "é", x))

# fix tildas
data['word'] = data['word'].apply(lambda x: re.sub("n~", "ñ", x))

# remove phones `sp` and `.sil`
data = data[(data['phone'] != ".sil") & (data['phone'] != "sp")]

data = data.reset_index(drop = True)

data.sample(25)

# fix naming of participant col
data["Participant"] = data["Participant"].apply(lambda x: x[:4])

data.to_csv("data/alldata.csv", index = False)

# remove rows not containing vowels
vowelsdf = data[(data['phone']=="a") | 
                 (data['phone']=="e") | 
                 (data['phone']=="i") | 
                 (data['phone']=="u") | 
                 (data['phone']=="o")]
vowelsdf = vowelsdf.reset_index(drop = True)
vowelsdf = vowelsdf.rename(columns = {"phone": "Vowel"})
len(vowelsdf)

vowelsdf.to_csv("data/allvowels.csv", index = False)
