==> 04_exploratory.ipynb <==
# Exploratory analyses

import pandas as pd

# import data
df = pd.read_csv("data/balancednormdata.csv")

First, calculate the median `F1.50_norm` and `F2.50_norm` by `Vowel` for each `Participant`.

# for creating a table in paper
df.groupby(["Corpus", "Gender", "Vowel"])[["F1.50_norm", "F2.50_norm"]].agg(["mean", "std"])

# for plotting in R
vowel_plot = df.groupby(["Corpus", "Gender", "Vowel"])[["F1.50_norm", "F2.50_norm"]].mean()
vowel_plot = vowel_plot.reset_index()
vowel_plot

vowel_plot.to_csv("plot_model/area_plot.csv", index = False)

# add duration_norm, duration_ms and duration_log cols
vowel_model = df.copy()
vowel_model["Dur_ms"] = vowel_model["dur_ph"].apply(lambda x: x*1000)
vowel_model["Dur_norm"] = vowel_model["Dur_ms"] * vowel_model["Speech Rate"]
vowel_model["Dur_log"] = vowel_model["Dur_ms"].apply(lambda x: math.log(x))

vowel_model.to_csv("plot_model/vowel_model.csv", index = False)

## Vowel space area by speaker

parts = df.groupby(["Participant", "Vowel"])[["F1.50_norm", "F2.50_norm"]].mean()
parts_dict = parts.groupby(level=0).apply(lambda parts: parts.xs(parts.name).to_dict()).to_dict()
parts_dict

# takes dicts in form {participant: {F1: {a: formant, e: formant, etc.}, F2: {a: etc.}}}
import math

def PolygonArea(d):
    areas = {}
    
    def area(a, b, c):
        def distance(p1, p2):
            return math.hypot(p1[0]-p2[0], p1[1]-p2[1])

        side_a = distance(a, b)
        side_b = distance(b, c)
        side_c = distance(c, a)
        s = 0.5 * ( side_a + side_b + side_c)
        return math.sqrt(s * (s - side_a) * (s - side_b) * (s - side_c))
    
    for p in d.keys():
        # coordinates in form (-F2, -F1)
        a = (d[p]["F2.50_norm"]["a"], d[p]["F1.50_norm"]["a"])
        e = (d[p]["F2.50_norm"]["e"], d[p]["F1.50_norm"]["e"])
        i = (d[p]["F2.50_norm"]["i"], d[p]["F1.50_norm"]["i"])
        o = (d[p]["F2.50_norm"]["o"], d[p]["F1.50_norm"]["o"])
        u = (d[p]["F2.50_norm"]["u"], d[p]["F1.50_norm"]["u"])
        
        a1 = area(a, e, o)
        a2 = area(e, o, u)
        a3 = area(i, e, u)
        
        areas[p] = a1 + a2 + a3
    return areas

parts_area = PolygonArea(parts_dict)
areas = pd.DataFrame.from_dict(parts_area, orient = "index").reset_index()
areas.columns = ["Participant", "Area"]
areas

# save into dataset
area_model = df.merge(areas, on = "Participant", how = "outer")
area_model.head(1)

area_model.to_csv("plot_model/area_model.csv", index = False)

## Add in vs centroid and category centroid by speaker

cent = df.groupby("Participant")[["F1.50_norm", "F2.50_norm"]].mean().reset_index()
cent.rename(columns={"F1.50_norm": "F1_mean", "F2.50_norm": "F2_mean"}, inplace = True)
cent

vowel_means = df.groupby(["Participant", "Vowel"])[["F1.50_norm", "F2.50_norm"]].mean().reset_index()
vowel_means.rename(columns={"F1.50_norm": "F1_v", "F2.50_norm": "F2_v"}, inplace = True)
vowel_means.head()

dispersion = vowel_means.merge(cent, on = "Participant", how = "outer")
dispersion.sample(5)

import numpy as np

dispersion['Dispersion_v'] = np.sqrt((abs(dispersion["F1_mean"] - dispersion["F1_v"])**2) + (abs(dispersion["F2_mean"] - dispersion["F2_v"])**2))
disperse = dispersion.groupby("Participant")["Dispersion_v"].mean().reset_index()
disperse.rename(columns = {"Dispersion_v": "Dispersion"}, inplace = True)
disperse

# save dispersion to dataset
disp_model = df.merge(disperse, on = "Participant", how = "outer")
disp_model.sample(5)

disp_model.to_csv("plot_model/disp_model.csv", index = False)

# now create new set for disp plotting
disp_plot1 = df.groupby(["Corpus", "Gender", "Vowel"])[["F1.50_norm", "F2.50_norm"]].mean().reset_index()
disp_plot1.rename(columns = {"F1.50_norm": "F1_v", "F2.50_norm": "F2_v"}, inplace = True)
disp_plot1

disp_plot2 = df.groupby(["Corpus", "Gender"])[["F1.50_norm", "F2.50_norm"]].mean().reset_index()
disp_plot2.rename(columns = {"F1.50_norm": "F1_cent", "F2.50_norm": "F2_cent"}, inplace = True)
disp_plot2

disp_plot = disp_plot1.merge(disp_plot2, on = ["Corpus", "Gender"], how = "outer")
disp_plot

# save disp_plot
disp_plot.to_csv("plot_model/disp_plot.csv", index = False)

## /a/ reduction across stress

# data set for plotting
a = df[df["Vowel"]=="a"].copy()
a_stress = a.groupby(["Corpus", "Gender", "stress"])[["F1.50_norm", "F2.50_norm"]].mean()
a_stress = a_stress.reset_index()
a_stress

# save for plotting
a_stress.to_csv("plot_model/a_stress_plot.csv", index = False)

# save for model
a.to_csv("plot_model/a_stress_model.csv", index = False)

## Consonantal context

cons_context = df[(df["prev_ph"]=="t") |
                  (df["prev_ph"]=="p") |
                  (df["prev_ph"]=="s")].copy()
len(cons_context)

cons_context.groupby(["Corpus", "prev_ph", "Vowel"]).count()

cons_context["Dur_ms"] = cons_context["dur_ph"].apply(lambda x: x*1000)
cons_context["Dur_norm"] = cons_context["Dur_ms"] * cons_context["Speech Rate"]
cons_context["Dur_log"] = cons_context["Dur_ms"].apply(lambda x: math.log(x))

# save for modeling
cons_context.to_csv("plot_model/cons_cont.csv", index = False)

# for plotting
cons_plot = cons_context.groupby(["Corpus", "Gender", "Vowel", "prev_ph"])[["F1.50_norm", "F2.50_norm", "dur_ph"]].mean()
cons_plot = cons_plot.reset_index()
cons_plot["Dur_ms"] = cons_plot["dur_ph"].apply(lambda x: x*1000)
cons_plot

# save for plotting
cons_plot.to_csv("plot_model/cons_cont_plot.csv", index = False)
