==> 05_a_stress.Rmd <==
---
title: "Analysis of /a/ production across Stress"
author: "Annie Helms (annie_helms@berkeley.edu)"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = "cairo_pdf")
options(scipen = 999) # avoid scientific notation
```

Library packages

```{r message = FALSE, warning = FALSE}
library(r2glmm)
library(dplyr)
library(ggplot2)
library(lmerTest) #mixed effects linear regression
library(ggsignif) # helps plot significant relations in ggplot2
library(kableExtra) # formats dataframes in html output
library(BSDA) # manual t-test with means and sd info
library(emmeans) # tukey post-hoc tests
```

### Upload data set

```{r message = FALSE, echo = FALSE}
setwd("/Users/Annie/Box Sync/PhD/github/Bay_Area_Spanish/")
```

Upload productions of /a/ across stress, corpus, and gender.

```{r}
stress_plot = read.csv("plot_model/a_stress_plot.csv", 
                  stringsAsFactors = TRUE)
head(stress_plot, 5) %>%
  kable() %>% # piped to allow for kable formatting
  kable_styling(bootstrap_options = c("striped", # alternates row color in table
                                      "hover", # rows are highlighted when hovered over
                                      "condensed", # row height is condensed
                                      "responsive")) %>% # horizontal scrolling across table
  scroll_box(width = "100%")
```

For plotting purposes, upload vowel plot.

```{r}
area_plot = read.csv("plot_model/area_plot.csv", 
                  stringsAsFactors = TRUE)
head(area_plot, 5) %>%
  kable() %>% # piped to allow for kable formatting
  kable_styling(bootstrap_options = c("striped", # alternates row color in table
                                      "hover", # rows are highlighted when hovered over
                                      "condensed", # row height is condensed
                                      "responsive")) %>% # horizontal scrolling across table
  scroll_box(width = "100%")
```

```{r}
# remove /a/ from the above set
area_plot_no_a = area_plot %>%
  filter(Vowel != "a")

head(area_plot_no_a, 5) %>%
  kable() %>% # piped to allow for kable formatting
  kable_styling(bootstrap_options = c("striped", # alternates row color in table
                                      "hover", # rows are highlighted when hovered over
                                      "condensed", # row height is condensed
                                      "responsive")) %>% # horizontal scrolling across table
  scroll_box(width = "100%")
```

```{r message = FALSE, echo = FALSE, warning = FALSE}
stress = ggplot(area_plot_no_a,
                   aes(x = F2.50_norm,
                       y = F1.50_norm,
                       shape = Corpus:Gender)) +
  geom_point(aes(color = Vowel),
                 size = 2,
             alpha = 0.6) +
  geom_point(data = stress_plot,
             aes(x = F2.50_norm,
                 y = F1.50_norm,
                 size = stress)) +
  geom_label(aes(x = 2.1,
                 y = 0.3,
                 label = "i"),
             color = "black") +
  geom_label(aes(x = 1.8,
                 y = 0.48,
                 label = "e"),
             color = "black") +
  geom_label(aes(x = 1.45,
                 y = 0.63,
                 label = "a"),
             color = "black") +
  geom_label(aes(x = 0.9,
                 y = 0.33,
                 label = "u"),
             color = "black") +
  geom_label(aes(x = 1.12,
                 y = 0.46,
                 label = "o"),
             color = "black") +
  theme_classic() +
  xlim(c(2.3, 0.8)) + # reverses axis scale
  ylim(c(0.72, 0.28)) + # reverses axis scale
  labs(x = "Normalized F2",
       y = "Normalized F1",
       title = "Spanish vowel space: /a/ across stress") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.position = "right") +
  guides(color = FALSE,
         size = guide_legend(title = "Stress"),
         shape = guide_legend(title = "Speaker Profile")) +
  scale_size_manual(values = c(2,3),
                    labels = c("Stressed", "Unstressed")) +
  scale_shape_manual(labels = c("CBAS Female",
                                "CBAS Male",
                                "DIMEx100 Female",
                                "DIMEx100 Male"),
                     values = c(15, 16, 0, 1))
stress
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/stress_a.pdf", family = "LM Roman 10", width = 7, height = 4)
stress
dev.off()
```

Now to see if the visual differences are statistically significant. Based on the plot above, only F1 will be examined.

```{r}
# import data set for modeling
vowels = read.csv("plot_model/vowel_model.csv", 
                  stringsAsFactors = TRUE)
stress_model = vowels %>% 
  filter(Vowel=="a")
```

```{r}
# submit to mixed effects linear regression model
stress_a_regression = lmer(F1.50_norm ~ Corpus * Gender * stress + (1|Participant),
                   data = stress_model, REML = FALSE)
summary(stress_a_regression)
```

Based on the model output above, the DIMEx100 speakers have higher productions of /a/ than CBAS speakers, and all speakers show centralization of /a/ in atonic syllables. Since the interaction of Corpus and Stress is not significant, it does not appear that Bay Area Spanish has any sort of exogenous variation (due to contact with California English).

```{r}
# submit to mixed effects linear regression model
stress_dur = lmer(Dur_norm ~ Corpus * Gender * stress + (1|Participant),
                   data = stress_model, REML = FALSE)
summary(stress_dur)
```

Post-hoc:

```{r message = FALSE}
emmeans(stress_dur, list(pairwise ~ Corpus * stress), adjust="tukey")
```

Based on the output of the post-hoc test above, within CBAS, normalized duration is longer in stressed vowels than in unstressed vowels (p < 0.001). Additionally, the duration of unstressed vowels is shorter in CBAS than in DIMEx100 (p < 0.001), which may be a contact effect from California English.

Plot this:

```{r message = FALSE, echo = FALSE, warning = FALSE}
# duration in ms
dur_stress_plot = ggplot(stress_model,
                   aes(x = Corpus:stress,
                       y = Dur_norm,
                       fill = Corpus:stress)) +
  geom_violin(draw_quantiles = c(.25, .50, .75)) +
  theme_classic() +
  scale_x_discrete(labels = c("CBAS stressed",
                              "CBAS unstressed",
                              "DIMEx100 stressed",
                              "DIMEx100 unstressed")) +
  geom_signif(comparisons = list(c("CBAS:stressed", "CBAS:unstressed")),
              y_position = 575,
              annotation = "***") +
  geom_signif(comparisons = list(c("CBAS:unstressed", "DIMEx100:unstressed")),
              y_position = 650,
              annotation = "***") +
  #xlim(c(2.2, 0.85)) + # reverses axis scale
  ylim(c(50, 700)) + # reverses axis scale
  labs(x = "Corpus and stress type",
       y = "Duration multiplied by speech rate",
       title = "Normalized duration of /a/ across corpus and stress") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12)) +
   guides(fill = FALSE)
  # scale_size_manual(values = c(3,4))
dur_stress_plot
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/dur_stress_plot.pdf", family = "LM Roman 10", width = 7, height = 5)
dur_stress_plot
dev.off()
```
