==> 06_categories_gender.Rmd <==
---
title: "Analysis of vowel categories across corpus and gender"
author: "Annie Helms (annie_helms@berkeley.edu)"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, dev = "cairo_pdf")
options(max.print = 1500, # prints full model output
        scipen = 999) # avoids scientific notation
```

Library packages

```{r message = FALSE, warning = FALSE}
library(r2glmm)
library(dplyr)
library(ggplot2)
library(lmerTest) #mixed effects linear regression
library(ggsignif) # helps plot significant relations in ggplot2
library(kableExtra) # formats dataframes in html output
library(BSDA) # manual t-test with means and sd info
library(emmeans) # tukey post-hoc tests
```

### Upload data set

```{r message = FALSE, echo = FALSE}
setwd("/Users/Annie/Box Sync/PhD/github/Bay_Area_Spanish/")
```

Upload productions of vowels across corpus and gender.

```{r}
vowel_plot = read.csv("plot_model/area_plot.csv", 
                  stringsAsFactors = TRUE)
head(vowel_plot, 5) %>%
  kable() %>% # piped to allow for kable formatting
  kable_styling(bootstrap_options = c("striped", # alternates row color in table
                                      "hover", # rows are highlighted when hovered over
                                      "condensed", # row height is condensed
                                      "responsive")) %>% # horizontal scrolling across table
  scroll_box(width = "100%")
```

```{r message = FALSE, echo = FALSE, warning = FALSE}
vs = ggplot(area_plot,
                   aes(x = F2.50_norm,
                       y = F1.50_norm,
                       shape = Corpus:Gender)) +
  geom_point(aes(color = Vowel),
                 size = 2) +
  geom_label(aes(x = 2.1,
                 y = 0.3,
                 label = "i"),
             color = "black") +
  geom_label(aes(x = 1.8,
                 y = 0.48,
                 label = "e"),
             color = "black") +
  geom_label(aes(x = 1.45,
                 y = 0.63,
                 label = "a"),
             color = "black") +
  geom_label(aes(x = 0.9,
                 y = 0.33,
                 label = "u"),
             color = "black") +
  geom_label(aes(x = 1.12,
                 y = 0.46,
                 label = "o"),
             color = "black") +
  theme_classic() +
  xlim(c(2.3, 0.8)) + # reverses axis scale
  ylim(c(0.72, 0.28)) + # reverses axis scale
  labs(x = "Normalized F2",
       y = "Normalized F1",
       title = "Spanish vowel space across corpus and gender") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.position = c(0.2, 0.25)) +
  guides(color = FALSE,
         shape = guide_legend(title = "Speaker Profile")) +
  scale_shape_manual(labels = c("CBAS Female",
                                "CBAS Male",
                                "DIMEx100 Female",
                                "DIMEx100 Male"),
                     values = c(15, 16, 0, 1))
vs
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/vs.pdf", family = "LM Roman 10", width = 7, height = 5)
vs
dev.off()
```

Below, the relationship between duration and (vowel category, corpus, and gender) will be plotted.

```{r}
# import data set: `vowel_model` which includes three different measures of duration
vowel_model = read.csv("plot_model/vowel_model.csv", 
                  stringsAsFactors = TRUE)
head(vowel_model, 5) %>%
  kable() %>% # piped to allow for kable formatting
  kable_styling(bootstrap_options = c("striped", # alternates row color in table
                                      "hover", # rows are highlighted when hovered over
                                      "condensed", # row height is condensed
                                      "responsive")) %>% # horizontal scrolling across table
  scroll_box(width = "100%")
```

Re-level vowel factor:

```{r}
vowel_model$Vowel = factor(vowel_model$Vowel, levels = c("i", "e", "a", "o", "u"))
```

First, calculate the average duration per vowel category per gender per corpus.

```{r}
duration_plots = vowel_model %>%
  group_by(Corpus, Gender, Vowel) %>%
  summarize(dur_ms = mean(Dur_ms),
            dur_norm = mean(Dur_norm),
            dur_log = mean(Dur_log),
            F2 = mean(F2.50_norm))
duration_plots
```

Now, to plot each measure of duration.

```{r message = FALSE, echo = FALSE, warning = FALSE}
# duration in ms
dur_ms_plot = ggplot(duration_plots,
                   aes(x = F2,
                       y = dur_ms,
                       linetype = Gender,
                       color = Corpus)) +
  geom_path(aes(group = Corpus:Gender)) +
  geom_label(aes(label = Vowel),
             color = "black") +
  theme_classic() +
  xlim(c(2.2, 0.85)) + # reverses axis scale
  ylim(c(130, 50)) + # reverses axis scale
  labs(x = "Normalized F2",
       y = "Duration (ms)",
       title = "Duration (ms) of vowels across corpus and gender") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.position = c(0.2,0.26),
        legend.box = "horizontal") +
  guides(shape = FALSE) +
  scale_size_manual(values = c(3,4))
dur_ms_plot
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/dur_ms_plot.pdf", family = "LM Roman 10", width = 7, height = 5)
dur_ms_plot
dev.off()
```

Now same thing but with normalized duration.

```{r message = FALSE, echo = FALSE, warning = FALSE}
# duration in ms
dur_norm_plot = ggplot(duration_plots,
                   aes(x = F2,
                       y = dur_norm,
                       linetype = Gender,
                       color = Corpus)) +
  geom_path(aes(group = Corpus:Gender)) +
  geom_label(aes(label = Vowel),
             color = "black") +
  theme_classic() +
  xlim(c(2.2, 0.85)) + # reverses axis scale
  ylim(c(190, 120)) + # reverses axis scale
  labs(x = "Normalized F2",
       y = "Duration multiplied by speech rate",
       title = "Normalized duration of Spanish vowels across corpus and gender") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.position = c(0.2,0.24),
        legend.box = "horizontal") +
  guides(shape = FALSE) +
  scale_size_manual(values = c(3,4))
dur_norm_plot
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/dur_norm_plot.pdf", family = "LM Roman 10", width = 7, height = 5)
dur_norm_plot
dev.off()
```

Same plot but with log duration.

```{r message = FALSE, echo = FALSE, warning = FALSE}
# duration in ms
dur_log_plot = ggplot(duration_plots,
                   aes(x = F2,
                       y = dur_log,
                       linetype = Gender,
                       color = Corpus)) +
  geom_path(aes(group = Corpus:Gender)) +
  geom_label(aes(label = Vowel),
             color = "black") +
  theme_classic() +
  xlim(c(2.2, 0.85)) + # reverses axis scale
  ylim(c(4.8, 3.9)) + # reverses axis scale
  labs(x = "Normalized F2",
       y = "Duration (log)",
       title = "Log duration of Spanish vowels across corpus and gender") +
  theme(plot.title = element_text(hjust = 0.5, size = 16),
        axis.title = element_text(size = 12),
        legend.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.position = c(0.2,0.24),
        legend.box = "horizontal") +
  guides(shape = FALSE) +
  scale_size_manual(values = c(3,4))
dur_log_plot
```

```{r echo = FALSE, eval = FALSE}
cairo_pdf("figs/dur_log_plot.pdf", family = "LM Roman 10", width = 7, height = 5)
dur_log_plot
dev.off()
```

Now to run some models. The DVs that will be used in mixed effects linear regression models are: F1, F2, Duration (ms), Duration (normalized) and Duration (log transformed).

First the three duration models:

```{r}
# submit to mixed effects linear regression model
dur_ms_model = lmer(Dur_ms ~ Corpus * Gender * Vowel + (1|Participant),
                   data = vowel_model, REML = FALSE)
summary(dur_ms_model)
```

This model makes the most sense to analyze as it controls for varying speech rates.

```{r}
# submit to mixed effects linear regression model
dur_norm_model = lmer(Dur_norm ~ Corpus * Gender * Vowel + (1|Participant),
                   data = vowel_model, REML = FALSE)
summary(dur_norm_model)
```

Run the appropriate post-hocs:

```{r message = FALSE}
emmeans(dur_norm_model, list(pairwise ~ Vowel), adjust="tukey")
```

According to the output of the post-hoc test above and below, both /a/ and /o/ have significantly longer durations than the high and mid vowels (p < 0.001 for all). 

Within corpus: This holds for CBAS, except that the duration of /a/ is not significantly longer than that of /u/ in CBAS, and the duration of /o/ is longer than that of /a/ in CBAS (p < 0.001 for both). In DIMEx100, /o/ is not different from either /e/ or /i/. Additionally, in DIMEx100, /a/ is longer than /o/.

Between corpora: /a/ is significantly longer in DIMEx100 than in CBAS, and /o/ is significantly longer in CBAS than in DIMEx100 (p < 0.001 for both).

```{r message = FALSE}
emmeans(dur_norm_model, list(pairwise ~ Vowel * Corpus), adjust="tukey")
```

```{r}
# submit to mixed effects linear regression model
dur_log_model = lmer(Dur_log ~ Corpus * Gender * Vowel + (1|Participant),
                   data = vowel_model, REML = FALSE)
summary(dur_log_model)
```

Now to model across F1 and F2.

```{r}
# submit to mixed effects linear regression model
f1_model = lmer(F1.50_norm ~ Corpus * Gender * Vowel + (1|Participant),
                   data = vowel_model, REML = FALSE)
summary(f1_model)
```

Appropriate post-hoc tests:

```{r message = FALSE}
emmeans(f1_model, list(pairwise ~ Vowel * Gender * Corpus), adjust="tukey")
```

According the the post-hoc output above:

Within corpora: /o/ is lower in female speech than in male speech in CBAS (p < 0.05). No gender differences appear in DIMEx100.

Between corpora: Among females, /i/, /e/, /a/, and /o/ are lower in CBAS than in DIMEx100 (p < 0.01, p < 0.001, p < 0.01, and p < 0.001). Among males, /i/ and /e/ are lower in CBAS than in DIMEx100 (p < 0.001 and p < 0.05).


```{r}
# submit to mixed effects linear regression model
f2_model = lmer(F2.50_norm ~ Corpus * Gender * Vowel + (1|Participant),
                   data = vowel_model, REML = FALSE)
summary(f2_model)
```

Appropriate post-hoc tests:

```{r message = FALSE}
emmeans(f2_model, list(pairwise ~ Vowel * Gender * Corpus), adjust="tukey")
```

According to the output of the post-hoc above:

Within corpora: In CBAS, females produce /i/ more backed than males (p < 0.001).

Between corpora: Among females, CBAS speakers produce /i/, /e/, /a/, and /o/ as more backed than DIMEx100 speakers (p < 0.01 for all). Among males, CBAS speakers produce /a/, /o/, and /u/ as more backed than DIMEx100 speakers (p < 0.001, p < 0.001, and p < 0.05).
