==> 09_ao_duration.ipynb <==
# Anomalous durations of /a/ and /o/

The universal tendency is that low vowels (here, /a/) will be produced with longer durations than high vowels. This prediction plays out in the DIMEx100 Corpus, but not in CBAS. In CBAS, /o/ is produced with the longest duration, followed by /a/, contrary to prediction.

import pandas as pd

vowels = pd.read_csv("plot_model/vowel_model.csv")
ao = vowels[(vowels["Vowel"]=="a") | (vowels["Vowel"]=="o")].copy()
len(ao)

ao.groupby(["Corpus", "Vowel", "stress"]).Dur_norm.mean()

It seems that the duration of /a/ and /o/ are equal, but that /o/ does not show an effect of stress as /a/ does, thus skewing the average to be higher than that of /a/. 

# now with raw duration
ao.groupby(["Corpus", "Vowel", "stress",]).Dur_ms.mean()

# Now check for following segment, where in English, duration is longest before a voiced consonant
# first group consonants according to voicing
ao.next_ph.unique()

# drop nan row
nan_df = ao[ao.isna().any(axis=1)]
nan_df[["word", "prev_ph", "next_ph"]]

# replace nan with ".sil"
ao.fillna(".sil", axis = 0, inplace = True)
len(ao)

ao.next_ph.unique()

import numpy as np

# create a list of our conditions
conditions = [
    ## first condition
    (ao["next_ph"] == "x") | (ao["next_ph"] == "k") | (ao["next_ph"] == "s") | (ao["next_ph"] == "p") |
    (ao["next_ph"] == "t") | (ao["next_ph"] == "f") | (ao["next_ph"] == "tS") | (ao["next_ph"] == "T"),
    ## second condition
    (ao["next_ph"] == "m") | (ao["next_ph"] == "l") | (ao["next_ph"] == "n") | (ao["next_ph"] == "rf") |
    (ao["next_ph"] == "b") | (ao["next_ph"] == "G") | (ao["next_ph"] == "D") | (ao["next_ph"] == "ng") |
    (ao["next_ph"] == "r") | (ao["next_ph"] == "d") | (ao["next_ph"] == "g") | (ao["next_ph"] == "r(") |
    (ao["next_ph"] == "n~") | (ao["next_ph"] == "Z") | (ao["next_ph"] == "L"),
    ## third condition
   (ao["next_ph"] == "sp") | (ao["next_ph"] == ".sil") | (ao["next_ph"] == ".bn") | (ao["next_ph"] == ".0")
    ]

# create a new column and use np.select to assign values to it using our lists as arguments
ao["voicing_next_ph"] = np.select(conditions, ["voiceless", "voiced", "other"])

ao.groupby(["Corpus", "Vowel", "stress", "voicing_next_ph"]).Dur_norm.mean()

# remove 0.bn and .0 contexts
ao_fixed = ao[(ao["prev_ph"]!=".bn") & (ao["next_ph"]!=".bn")
             & (ao["next_ph"]!=".0") & (ao["next_ph"]!=".0")]
len(ao_fixed)

ao_fixed.groupby(["Corpus", "Vowel", "stress"]).Dur_norm.mean()

#remove word final context to avoid duration calculation issues from background periodicity
ao_fixed_no_fin = ao[ao["voicing_next_ph"]!= "other"]
len(ao_fixed_no_fin)

ao_fixed_no_fin.groupby(["Corpus", "Vowel"]).Dur_norm.mean()

ao_fixed.groupby(["Corpus", "Vowel", "is_wdfin_ph"]).Gender.count()

ao_fixed.groupby(["Corpus", "Vowel", "is_wdfin_ph"]).Dur_norm.mean()

# CBAS female production of /i/

i = vowels[vowels["Vowel"]=="i"].copy()
len(i)

i.groupby(["Corpus", "Gender"]).stress.count()

i.groupby(["Corpus", "Gender", "stress"]).Participant.count()

i.groupby(["Corpus", "Gender", "stress"])["F2.50_norm"].mean()

# Is it all females or just one?
i_fem_cbas = i[(i["Gender"]=="Female") & (i["Corpus"]=="CBAS")].copy()
i_fem_cbas.groupby(["Participant", "stress"])["F2.50_norm"].mean()

For some female CBAS speakers, stress seems to play a large role in /i/ F2 production, where unstressed /i/ is more backed (more central). However, most of the backing of /i/ seen globally is due to the productions from just one participant, p115.

# check if p115's productions for all vowels is different

fem_cbas = vowels[(vowels["Gender"]=="Female") & (vowels["Corpus"]=="CBAS")].copy()
fem_cbas.groupby(["Vowel", "Participant"])[["F1.50_norm", "F2.50_norm"]].mean()
